+++ src/torchoutil/nn/modules/crop.py
@@ -5,11 +5,15 @@ from typing import Iterable, Union
 
 from torch import Generator, Tensor, nn
 
-from torchoutil.nn.functional.crop import crop_dim, crop_dims, CropAlign
+from torchoutil.nn.functional.crop import CropAlign, crop_dim, crop_dims
 from torchoutil.utils.collections import dump_dict
 
 
 class CropDim(nn.Module):
+    """
+    For more information, see :func:`~torchoutil.nn.functional.crop.crop_dim`.
+    """
+
     def __init__(
         self,
         target_length: int,
@@ -46,6 +50,10 @@ class CropDim(nn.Module):
 
 
 class CropDims(nn.Module):
+    """
+    For more information, see :func:`~torchoutil.nn.functional.crop.crop_dims`.
+    """
+
     def __init__(
         self,
         target_lengths: Iterable[int],