+++ src/torchoutil/nn/modules/multiclass.py
@@ -23,6 +23,10 @@ T = TypeVar("T")
 
 
 class IndicesToOneHot(nn.Module):
+    """
+    For more information, see :func:`~torchoutil.nn.functional.multiclass.indices_to_onehot`.
+    """
+
     def __init__(
         self,
         num_classes: int,
@@ -63,6 +67,10 @@ class IndicesToOneHot(nn.Module):
 
 
 class IndicesToNames(nn.Module, Generic[T]):
+    """
+    For more information, see :func:`~torchoutil.nn.functional.multiclass.indices_to_names`.
+    """
+
     def __init__(
         self,
         idx_to_name: Mapping[int, T],
@@ -79,6 +87,10 @@ class IndicesToNames(nn.Module, Generic[T]):
 
 
 class OneHotToIndices(nn.Module):
+    """
+    For more information, see :func:`~torchoutil.nn.functional.multiclass.onehot_to_indices`.
+    """
+
     def __init__(self) -> None:
         super().__init__()
 
@@ -91,6 +103,10 @@ class OneHotToIndices(nn.Module):
 
 
 class OneHotToNames(nn.Module, Generic[T]):
+    """
+    For more information, see :func:`~torchoutil.nn.functional.multiclass.onehot_to_names`.
+    """
+
     def __init__(
         self,
         idx_to_name: Mapping[int, T],
@@ -107,6 +123,10 @@ class OneHotToNames(nn.Module, Generic[T]):
 
 
 class NamesToIndices(nn.Module, Generic[T]):
+    """
+    For more information, see :func:`~torchoutil.nn.functional.multiclass.names_to_indices`.
+    """
+
     def __init__(
         self,
         idx_to_name: Mapping[int, T],
@@ -123,6 +143,10 @@ class NamesToIndices(nn.Module, Generic[T]):
 
 
 class NamesToOneHot(nn.Module, Generic[T]):
+    """
+    For more information, see :func:`~torchoutil.nn.functional.multiclass.names_to_onehot`.
+    """
+
     def __init__(
         self,
         idx_to_name: Mapping[int, T],
@@ -158,6 +182,10 @@ class NamesToOneHot(nn.Module, Generic[T]):
 
 
 class ProbsToIndices(nn.Module):
+    """
+    For more information, see :func:`~torchoutil.nn.functional.multiclass.probs_to_indices`.
+    """
+
     def forward(
         self,
         probs: Tensor,
@@ -167,6 +195,10 @@ class ProbsToIndices(nn.Module):
 
 
 class ProbsToOneHot(nn.Module):
+    """
+    For more information, see :func:`~torchoutil.nn.functional.multiclass.probs_to_onehot`.
+    """
+
     def __init__(
         self,
         *,
@@ -199,6 +231,10 @@ class ProbsToOneHot(nn.Module):
 
 
 class ProbsToNames(nn.Module, Generic[T]):
+    """
+    For more information, see :func:`~torchoutil.nn.functional.multiclass.probs_to_names`.
+    """
+
     def __init__(
         self,
         idx_to_name: Mapping[int, T],